small package refactor and add animations
