+++ app/src/main/java/com/ferdinand/pdftestapp/ui/composables/ErrorDialog.kt
@@ -1,4 +1,4 @@
-package com.ferdinand.pdftestapp.ui.main
+package com.ferdinand.pdftestapp.ui.composables
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
