+++ app/src/main/res/navigation/main_graph.xml
@@ -16,7 +16,11 @@
             app:popExitAnim="@anim/to_left" />
         <action
             android:id="@+id/action_pdfListFragment_to_searchFragment"
-            app:destination="@id/searchFragment" />
+            app:destination="@id/searchFragment"
+            app:enterAnim="@anim/from_left"
+            app:exitAnim="@anim/to_right"
+            app:popEnterAnim="@anim/from_right"
+            app:popExitAnim="@anim/to_left"/>
     </fragment>
 
     <fragment
@@ -24,7 +28,11 @@
         android:name="com.ferdinand.pdftestapp.ui.search.SearchFragment">
         <action
             android:id="@+id/action_searchFragment_to_pdfViewActivity"
-            app:destination="@id/pdfViewActivity" />
+            app:destination="@id/pdfViewActivity"
+            app:enterAnim="@anim/from_left"
+            app:exitAnim="@anim/to_right"
+            app:popEnterAnim="@anim/from_right"
+            app:popExitAnim="@anim/to_left" />
     </fragment>
 
     <activity
